# app/Http/Controllers/Controller.php
<?php

namespace App\Http\Controllers;

use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
use Illuminate\Foundation\Validation\ValidatesRequests;
use Illuminate\Routing\Controller as BaseController;

/**
     * @OA\Info(
     *      version="1.0.0",
     *      title="Blogs Api Documentation",
     *      description="L5 Swagger OpenApi description",
     *      @OA\Contact(
     *          email="admin@admin.com"
     *      ),
     *      @OA\License(
     *          name="Apache 2.0",
     *          url="http://www.apache.org/licenses/LICENSE-2.0.html"
     *      )
     * )
     *
     * @OA\Server(
     *      url=L5_SWAGGER_CONST_HOST,
     *      description="Demo API Server"
     * )
     *
     * @OA\Tag(
     *     name="Authentication",
     *     description="API Endpoints of Projects"
     * )
     * @OA\Tag(
     *     name="Category",
     *     description="API Endpoints of Projects"
     * )
     * @OA\Tag(
     *     name="Post",
     *     description="API Endpoints of Projects"
     * )
     * @OA\Tag(
     *     name="Favorite",
     *     description="API Endpoints of Projects"
     * )
     * @OA\Tag(
     *     name="Comment",
     *     description="API Endpoints of Projects"
     * )
     */

class Controller extends BaseController
{
    use AuthorizesRequests, ValidatesRequests;
}